+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Dependencies
 
+- Dropped support for Python 3.10
 - Bumped CBottle (and Earth2Grid) versions
 
 ## [0.9.0] - 2025-08-19